+++ app/src/main/java/com/myproject/campusmap_cleanarchitecture/data/datasource/localdatasource/BuildingDataSource.kt
@@ -11,7 +11,6 @@ import com.google.firebase.storage.StorageReference
 import com.myproject.campusmap_cleanarchitecture.R
 import com.myproject.campusmap_cleanarchitecture.data.db.local.database.BuildingDatabase
 import com.myproject.campusmap_cleanarchitecture.data.db.local.entity.BuildingEntity
-import com.myproject.campusmap_cleanarchitecture.domain.NetworkErrorHandler
 import javax.inject.Inject
 
 class BuildingDataSource @Inject constructor(
