+++ app/src/main/java/com/myproject/campusmap_cleanarchitecture/domain/error/NetworkError.kt
@@ -1,4 +1,4 @@
-package com.myproject.campusmap_cleanarchitecture.domain
+package com.myproject.campusmap_cleanarchitecture.domain.error
 
 sealed class NetworkError {
     object Unknown : NetworkError()
