+++ app/src/main/java/com/myproject/campusmap_cleanarchitecture/domain/usecase/GetBusPositionDataUseCase.kt
@@ -1,6 +1,6 @@
 package com.myproject.campusmap_cleanarchitecture.domain.usecase
 
-import com.myproject.campusmap_cleanarchitecture.domain.NetworkResult
+import com.myproject.campusmap_cleanarchitecture.domain.error.NetworkResult
 import com.myproject.campusmap_cleanarchitecture.domain.model.BusPosition
 import com.myproject.campusmap_cleanarchitecture.domain.repository.BusRepository
 import javax.inject.Inject
