+++ app/src/main/java/com/myproject/campusmap_cleanarchitecture/ui/lectureroom/lectureroomdetail/LectureRoomDetailViewModel.kt
@@ -14,6 +14,8 @@ class LectureRoomDetailViewModel @Inject constructor(
     private val getLectureRoomDetailImagesUseCase: GetLectureRoomDetailImagesUseCase
 ) : ViewModel(){
 
+
+
     fun getEngFirstLectureRoomDetail(id: Int) = getLectureRoomUseCase.invoke(id)
 
     fun getImagesUseCase(c: Context, path: String?, v: ImageView) = getLectureRoomDetailImagesUseCase.invoke(c, path, v)
